+++ src/cloud/mod.rs
@@ -1,8 +1,9 @@
 // CLASSIFICATION: COMMUNITY
-// Filename: mod.rs v0.1
+// Filename: mod.rs v0.2
 // Author: Lukas Bower
-// Date Modified: 2025-06-17
+// Date Modified: 2026-10-10
 
 //! Cloud-facing utilities.
 
 pub mod hooks;
+pub mod orchestrator;